split up gameplay.c into multiple files
