+++ src/drawing.h
@@ -3,7 +3,7 @@
 
 // #include "shared.h"
 #include "files.h"
-#include "gameplay.h"
+// #include "gameplay.h"
 
 #include "../deps/raylib/src/raylib.h"
