+++ src/gameplay/playing.h
@@ -1,0 +1,7 @@
+#include <stdbool.h>
+
+
+void fCountDown(bool reset);
+void fPlaying(bool reset);
+void fFail(bool reset);
+void fEndScreen(bool reset);
