+++ src/gameplay/recording.h
@@ -1,0 +1,1 @@
+void fRecording(bool reset);
